@@ -1,7 +1,9 @@
-// 路径: lib/repositories/analytics_repository.dart
+// 路径: (供应商 App) lib/repositories/analytics_repository.dart
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import '../models/order_model.dart'; // 确保 import 路径正确
+// --- ( ✨ 1. 更改导入 ✨ ) ---
+// 我们不再需要 OrderModel, 我们需要 ReviewModel
+import '../models/review.dart'; // <-- 确保你导入了 review.dart
 
 class AnalyticsRepository {
   final FirebaseAuth _auth = FirebaseAuth.instance;
@@ -9,24 +11,23 @@ class AnalyticsRepository {
 
   String? get _vendorId => _auth.currentUser?.uid;
 
-  // 获取所有 "已完成" 且 "已评价" 的订单
-  Stream<List<OrderModel>> getVendorReviewsStream() {
+  // --- ( ✨ 2. 彻底替换这个函数 ✨ ) ---
+  // (它现在返回 List<ReviewModel>)
+  Stream<List<ReviewModel>> getVendorReviewsStream() {
     final vendorId = _vendorId;
     if (vendorId == null) {
       throw Exception('User not logged in');
     }
 
     return _db
-        .collection('orders')
-        .where('vendorIds', arrayContains: vendorId)
-        .where('status', isEqualTo: 'Completed')
-        .where('rating', isGreaterThan: 0) // 假设所有评分都 > 0
-        .orderBy('rating', descending: true) // 先按评分排
-        .orderBy('reviewTimestamp', descending: true) // 再按时间排
+        .collection('reviews') // <-- ( ✨ 更改: 查询 'reviews' 集合 )
+        .where('vendorId', isEqualTo: vendorId) // <-- ( ✨ 更改: 按 vendorId 过滤 )
+        .orderBy('timestamp', descending: true) // <-- ( ✨ 更改: 按时间戳排序 )
         .snapshots()
         .map((snapshot) {
       return snapshot.docs
-          .map((doc) => OrderModel.fromMap(doc.data()!, doc.id))
+          .map((doc) => ReviewModel.fromMap(
+              doc.data())) // <-- ( ✨ 更改: 使用 ReviewModel.fromMap )
           .toList();
     });
   }
