@@ -8,11 +8,9 @@ class OrderRepository {
   final FirebaseAuth _auth = FirebaseAuth.instance;
   final FirebaseFirestore _db = FirebaseFirestore.instance;
 
-  // 获取当前供应商的 UID
   String? get _vendorId => _auth.currentUser?.uid;
 
-  // --- 1. 获取订单列表 (用于 Orders Page) ---
-  // ( ✨ 已包含你对 "paid pending pickup" 状态的修复 ✨ )
+  // --- ( 这是你现有的函数 - 保持不变 ) ---
   Stream<List<OrderModel>> getOrdersStream(String orderType) {
     final vendorId = _vendorId;
     if (vendorId == null) {
@@ -21,14 +19,23 @@ class OrderRepository {
 
     return _db
         .collection('orders')
-        .where('vendorIds', arrayContains: vendorId) // ( ❗ 依赖客户 App 修复 ❗ )
-        .where('orderType', isEqualTo: orderType)
+        .where('vendorIds', arrayContains: vendorId) // <-- ( ✨ 已修复! ✨ )
+        .where('orderType', isEqualTo: orderType) // <-- ( ✨ 正确! ✨ )
+
+        // --- ( ✨ 关键修复 ✨ ) ---
+        // 添加了 'Prepared' 状态
         .where('status', whereIn: [
-          'received',
+          'Received',
           'Preparing',
+          'Prepared', // <-- ( ✨ 新增状态 ✨ )
           'Ready for Pickup',
           'Delivering',
-          'paid pending pickup' // ( ✨ 已添加 ✨ )
+          'paid_pending_pickup',
+          'Pending Payment', // <-- NEW STATUS
+          'Payment Rejected', // <-- NEW STATUS
+          'Delivered',
+          'Picked Up'
+              'Cancelled' // <-- ( ✨ NEWLY ADDED ✨ )
         ])
         .orderBy('timestamp', descending: true)
         .snapshots()
@@ -43,8 +50,7 @@ class OrderRepository {
         });
   }
 
-  // --- 2. 更新订单状态 (用于 Details Page) ---
-  // ( ✨ 这是你缺失的函数之一 ✨ )
+  // --- ( 这是你现有的函数 - 保持不变 ) ---
   Future<void> updateOrderStatus(String orderId, String newStatus) async {
     try {
       await _db.collection('orders').doc(orderId).update({'status': newStatus});
@@ -54,8 +60,7 @@ class OrderRepository {
     }
   }
 
-  // --- 3. 分配司机 (用于 Details Page) ---
-  // ( ✨ 这是你缺失的函数之二 ✨ )
+  // --- ( 这是你现有的函数 - 保持不变 ) ---
   Future<void> assignDriverToOrder(String orderId) async {
     try {
       final driversSnapshot = await _db.collection('drivers').get();
@@ -80,23 +85,36 @@ class OrderRepository {
     }
   }
 
-  // --- 4. 获取今日统计 (用于 Home Page Dashboard) ---
-  // ( ✨ 这是你缺失的函数之三 ✨ )
+  // --- ( 这是你现有的函数 - 保持不变 ) ---
   Stream<Map<String, dynamic>> getTodaysStatsStream() {
     final vendorId = _vendorId;
     if (vendorId == null) {
       throw Exception('User not logged in');
     }
 
-    final now = DateTime.now();
-    final startOfToday = DateTime(now.year, now.month, now.day);
-    final startOfTodayTimestamp = Timestamp.fromDate(startOfToday);
+    // --- ( ✨ 关键修复：时区 ✨ ) ---
+    // 1. 获取当前的 UTC 时间
+    final nowUtc = DateTime.now().toUtc();
 
+    // 2. 手动将当前 UTC 时间调整为 UTC+8
+    final nowInMalaysia = nowUtc.add(const Duration(hours: 8));
+
+    // 3. 计算 UTC+8 时区的 "今天凌晨"
+    final startOfTodayInMalaysia = DateTime.utc(
+        nowInMalaysia.year, nowInMalaysia.month, nowInMalaysia.day, 0, 0, 0);
+    // 4. 将 UTC+8 的凌晨时间转换回 UTC 时间戳进行查询
+    final startOfTodayTimestamp = Timestamp.fromDate(
+        startOfTodayInMalaysia.subtract(const Duration(hours: 8)));
+    // --- ( ✨ 结束修复 ✨ ) ---
+
+    // 5. 查询 (这个查询现在是正确的)
+    // (它在查询: ...where('timestamp', >= '2025-11-14 16:00:00 UTC') )
     final query = _db
         .collection('orders')
-        .where('vendorIds', arrayContains: vendorId) // ( ❗ 依赖客户 App 修复 ❗ )
+        .where('vendorIds', arrayContains: vendorId)
         .where('timestamp', isGreaterThanOrEqualTo: startOfTodayTimestamp);
 
+    // 6. 监听快照 (这个逻辑是我们之前修复过的，是安全的)
     return query.snapshots().map((snapshot) {
       int orderCount = snapshot.docs.length;
       double totalSales = 0.0;
@@ -104,7 +122,10 @@ class OrderRepository {
       for (var doc in snapshot.docs) {
         final data = doc.data() as Map<String, dynamic>?;
         if (data != null && data.containsKey('total')) {
-          totalSales += (data['total'] as num?)?.toDouble() ?? 0.0;
+          final totalValue = data['total'];
+          if (totalValue is num) {
+            totalSales += totalValue.toDouble();
+          }
         }
       }
 
@@ -114,4 +135,72 @@ class OrderRepository {
       };
     });
   }
+  Stream<List<OrderModel>> getHistoryOrdersStream() {
+    final vendorId = _vendorId;
+    if (vendorId == null) {
+      throw Exception('User not logged in');
+    }
+
+    return _db
+        .collection('orders')
+        .where('vendorIds', arrayContains: vendorId)
+        // ( 关键 ) 查询 "Completed" 和 "Cancelled" 状态
+        .where('status', whereIn: ['Completed', 'Cancelled'])
+        // ( 索引 ) 这个查询需要一个新的索引
+        .orderBy('timestamp', descending: true)
+        .limit(50) // (可选：对历史记录进行分页或限制)
+        .snapshots()
+        .map((snapshot) {
+          return snapshot.docs.map((doc) {
+            final Map<String, dynamic>? data = doc.data();
+            if (data == null) {
+              throw Exception('Found order with empty data: ${doc.id}');
+            }
+            return OrderModel.fromMap(data, doc.id);
+          }).toList();
+        });
+  }
+
+  // --- ( ✨✨✨ IMPORTANT ✨✨✨ ) ---
+  //
+  // This function should ONLY be called by your Admin Panel or
+  // a Firebase Cloud Function *after* an Admin approves the payment.
+  //
+  // DO NOT call this from the customer app.
+  //
+  Future<void> deductStockForOrder(OrderModel order) async {
+    // Use a transaction to ensure this is atomic (all or nothing)
+    await _db.runTransaction((transaction) async {
+      for (final item in order.items) {
+        // 1. Get the reference to the product document
+        final productRef = _db
+            .collection('vendors')
+            .doc(item.vendorId)
+            .collection('products')
+            .doc(item.productId);
+
+        // 2. Get the product document *within the transaction*
+        final productSnapshot = await transaction.get(productRef);
+
+        if (!productSnapshot.exists) {
+          throw Exception(
+              'Product with ID ${item.productId} not found for vendor ${item.vendorId}.');
+        }
+
+        // 3. Get the current stock ('quantity' field in product.dart)
+        final currentStock = (productSnapshot.data()?['quantity'] as num?) ?? 0;
+
+        // 4. Calculate the new stock
+        final newStock = currentStock - item.quantity;
+
+        // 5. Update the product document *within the transaction*
+        // Set new stock, ensuring it doesn't go below 0.
+        transaction.update(productRef, {
+          'quantity': newStock < 0 ? 0 : newStock,
+        });
+      }
+    });
+    print('Stock successfully deducted for order ${order.id}');
+  }
+  // --- ( ✨✨✨ END OF CHANGE ✨✨✨ ) ---
 }
