@@ -1,13 +1,16 @@
+// 路径: lib/pages/vendor_home/account_settings_page.dart
 import 'package:flutter/material.dart';
 import '../../util/styles.dart';
 
 import 'edit_profile_page.dart';
 import 'edit_store_details_page.dart';
-// --- 1. 导入 Repository ---
+import 'edit_store_hours_page.dart';
+import 'delete_account_page.dart';
+
 import '../../repositories/vendor_data_repository.dart';
 
 class AccountSettingsPage extends StatefulWidget {
-  final VendorDataBundle bundle; // <-- 2. 接收数据
+  final VendorDataBundle bundle;
   const AccountSettingsPage({super.key, required this.bundle});
 
   @override
@@ -53,11 +56,40 @@ class _AccountSettingsPageState extends State<AccountSettingsPage> {
     );
   }
 
+  // --- (辅助) 危险操作的 Tile ---
+  Widget _buildDangerTile(String title, String subtitle, VoidCallback onTap) {
+    return Container(
+      margin: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 4.0),
+      decoration: BoxDecoration(
+        color: kCardColor,
+        borderRadius: BorderRadius.circular(12.0),
+        border:
+            Border.all(color: kPrimaryActionColor.withAlpha(70), width: 1.5),
+      ),
+      child: ListTile(
+        title: Text(title,
+            style: const TextStyle(
+                color: kPrimaryActionColor, fontWeight: FontWeight.bold)),
+        subtitle: Text(subtitle,
+            maxLines: 2,
+            overflow: TextOverflow.ellipsis,
+            style: TextStyle(color: kPrimaryActionColor.withAlpha(180))),
+        trailing: const Icon(Icons.chevron_right,
+            color: kPrimaryActionColor, size: 20),
+        onTap: onTap,
+      ),
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
-    // --- 4. (新增) 从 widget 中获取真实数据 ---
     final user = widget.bundle.user;
     final vendor = widget.bundle.vendor;
+
+    // 为营业时间创建一个可读的字符串
+    final String storeHoursSubtitle = vendor.storeHours.isEmpty
+        ? 'Tap to set your store hours'
+        : vendor.storeHours.join('\n'); // (e.g., "Mon: 9-5", "Tue: 9-5")
 
     return Scaffold(
       backgroundColor: kAppBackgroundColor,
@@ -71,28 +103,26 @@ class _AccountSettingsPageState extends State<AccountSettingsPage> {
       ),
       body: ListView(
         children: [
-          // --- 5. (已修改) 个人账户部分 ---
+          // --- 个人账户部分 ---
           _buildSectionHeader('ACCOUNT INFORMATION'),
           _buildInfoTile(
             'Full Name',
-            "${user.firstName} ${user.lastName}", // 真实数据
+            "${user.firstName} ${user.lastName}",
             () {
               Navigator.push(
                   context,
                   MaterialPageRoute(
-                      // --- 传递数据 ---
                       builder: (context) =>
                           EditProfilePage(bundle: widget.bundle)));
             },
           ),
           _buildInfoTile(
             'Email',
-            user.email, // 真实数据
+            user.email,
             () {
               Navigator.push(
                   context,
                   MaterialPageRoute(
-                      // --- 传递数据 ---
                       builder: (context) =>
                           EditProfilePage(bundle: widget.bundle)));
             },
@@ -104,40 +134,50 @@ class _AccountSettingsPageState extends State<AccountSettingsPage> {
               Navigator.push(
                   context,
                   MaterialPageRoute(
-                      // --- 传递数据 ---
                       builder: (context) =>
                           EditProfilePage(bundle: widget.bundle)));
             },
           ),
 
-          // --- 6. (已修改) 店铺设置部分 ---
+          // --- 店铺设置部分 ---
           _buildSectionHeader('STORE INFORMATION'),
           _buildInfoTile(
-            'Store Details', // <-- (修改) 移除 Hours
+            'Store Details',
             '${vendor.storeName}\n${vendor.storeAddress}',
             () {
               Navigator.push(
                   context,
                   MaterialPageRoute(
-                      // --- 传递数据 ---
                       builder: (context) =>
                           EditStoreDetailsPage(bundle: widget.bundle)));
             },
           ),
           _buildInfoTile(
-            'Payment Methods',
-            'Bank Transfer, E-Wallets',
+            'Store Hours',
+            storeHoursSubtitle,
             () {
-              // TODO: 跳转到 Payment Methods 页面
+              Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      // (传递 vendor 数据，因为它包含了 storeHours)
+                      builder: (context) =>
+                          EditStoreHoursPage(vendor: widget.bundle.vendor)));
             },
           ),
-          _buildInfoTile(
-            'Data & Backup',
-            'Manage your data',
+
+          // --- 危险区域 ---
+          _buildSectionHeader('DANGER ZONE'),
+          _buildDangerTile(
+            'Delete Account',
+            'Permanently delete your vendor account',
             () {
-              // TODO: 跳转到 Data & Backup 页面
+              Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => const DeleteAccountPage()));
             },
           ),
+          const SizedBox(height: 40),
         ],
       ),
     );
