@@ -2,12 +2,13 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import '../../util/styles.dart';
-// --- ( ✨ 关键修复 ✨ ) ---
-// 你很可能丢失了下面这两行 import
-import '../../models/order_model.dart';
-import '../../repositories/analytics_repository.dart'; // <-- 这一行修复了你的错误
-import 'order_review_details_page.dart';
-// --- ( ✨ 结束修复 ✨ ) ---
+// --- ( ✨ 1. 更改导入 ✨ ) ---
+// 导入 ReviewModel，不再需要 OrderModel 或 OrderReviewDetailsPage
+import '../../models/review.dart'; // <-- 导入 review.dart
+import '../../repositories/analytics_repository.dart';
+// (我们不再需要这个了)
+// import 'order_review_details_page.dart';
+// --- ( ✨ 结束修改 ✨ ) ---
 
 class AnalyticsPage extends StatefulWidget {
   const AnalyticsPage({super.key});
@@ -17,8 +18,10 @@ class AnalyticsPage extends StatefulWidget {
 }
 
 class _AnalyticsPageState extends State<AnalyticsPage> {
-  final AnalyticsRepository _repo = AnalyticsRepository(); // <-- 现在这一行不会报错了
-  late Stream<List<OrderModel>> _reviewsStream;
+  final AnalyticsRepository _repo = AnalyticsRepository();
+  // --- ( ✨ 2. 更改 Stream 的类型 ✨ ) ---
+  late Stream<List<ReviewModel>>
+      _reviewsStream; // <-- 从 OrderModel 改为 ReviewModel
 
   @override
   void initState() {
@@ -26,10 +29,10 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
     _reviewsStream = _repo.getVendorReviewsStream();
   }
 
-  // (辅助 Widget) 构建总评分卡
-  Widget _buildOverallRatingCard(List<OrderModel> orders) {
-    if (orders.isEmpty) {
-      // ( ✨ 新增 ✨ ) 如果没有评价，显示一个友好的卡片
+  // --- ( ✨ 3. 修改这个函数以接收 List<ReviewModel> ✨ ) ---
+  Widget _buildOverallRatingCard(List<ReviewModel> reviews) {
+    // <-- 类型已更改
+    if (reviews.isEmpty) {
       return Container(
         padding: const EdgeInsets.all(24.0),
         decoration: BoxDecoration(
@@ -44,13 +47,11 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
 
     // 1. 计算平均分
     double totalRating = 0;
-    for (var order in orders) {
-      // ( ✨ 修复 ✨ ) 确保只计算有评分的
-      if (order.rating != null) {
-        totalRating += order.rating!;
-      }
+    for (var review in reviews) {
+      // <-- 遍历 reviews
+      totalRating += review.rating; // <-- 使用 review.rating
     }
-    double avgRating = totalRating / orders.length;
+    double avgRating = totalRating / reviews.length;
 
     return Container(
       padding: const EdgeInsets.all(24.0),
@@ -87,7 +88,7 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
           _buildStarRating(avgRating, size: 28),
           const SizedBox(height: 8),
           Text(
-            'Based on ${orders.length} reviews',
+            'Based on ${reviews.length} reviews',
             style: TextStyle(
               fontSize: 14,
               color: kTextColor.withAlpha(150),
@@ -98,8 +99,9 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
     );
   }
 
-  // (辅助 Widget) 构建单个订单评价卡
-  Widget _buildReviewCard(OrderModel order) {
+  // --- ( ✨ 4. 修改这个函数以接收 ReviewModel ✨ ) ---
+  Widget _buildReviewCard(ReviewModel review) {
+    // <-- 类型已更改
     return Container(
       margin: const EdgeInsets.only(bottom: 12.0),
       decoration: BoxDecoration(
@@ -111,7 +113,8 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
         contentPadding:
             const EdgeInsets.symmetric(vertical: 12.0, horizontal: 16.0),
         title: Text(
-          'Order #${order.id.substring(0, 6)}...',
+          // ( ✨ 更改: 显示 Order ID )
+          'Order #${review.orderId.substring(0, 6)}...', // <-- 使用 review.orderId
           style: const TextStyle(
             color: kTextColor,
             fontSize: 16,
@@ -122,30 +125,28 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             const SizedBox(height: 4),
-            _buildStarRating(order.rating ?? 0),
+            _buildStarRating(review.rating), // <-- 使用 review.rating
             const SizedBox(height: 8),
             Text(
-              order.reviewText ?? '(No review text provided)',
+              review.reviewText.isEmpty // <-- 使用 review.reviewText
+                  ? '(No review text provided)'
+                  : review.reviewText,
               style: TextStyle(color: kTextColor.withAlpha(200), fontSize: 14),
               maxLines: 2,
               overflow: TextOverflow.ellipsis,
             ),
           ],
         ),
-        trailing: const Icon(Icons.chevron_right, color: kTextColor),
-        onTap: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => OrderReviewDetailsPage(order: order),
-            ),
-          );
-        },
+        // ( ✨ 更改: 移除了 onTap 和箭头，因为我们不再需要订单详情 )
+        // trailing: const Icon(Icons.chevron_right, color: kTextColor),
+        // onTap: () {
+        //   (已移除)
+        // },
       ),
     );
   }
 
-  // (辅助 Widget) 构建星级
+  // (辅助 Widget) 构建星级 (这个函数不需要修改)
   Widget _buildStarRating(double rating, {double size = 18}) {
     List<Widget> stars = [];
     for (int i = 1; i <= 5; i++) {
@@ -174,7 +175,9 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
           onPressed: () => Navigator.of(context).pop(),
         ),
       ),
-      body: StreamBuilder<List<OrderModel>>(
+      // --- ( ✨ 5. 更改 StreamBuilder 的类型 ✨ ) ---
+      body: StreamBuilder<List<ReviewModel>>(
+        // <-- 从 OrderModel 改为 ReviewModel
         stream: _reviewsStream,
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.waiting) {
@@ -182,6 +185,8 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
                 child: CircularProgressIndicator(color: kPrimaryActionColor));
           }
           if (snapshot.hasError) {
+            // ( ✨ 提示: 如果你在这里看到 "FAILED_PRECONDITION" 错误, )
+            // ( ✨ 那就是 Firebase 在要求你创建新索引 )
             return Center(
                 child: Padding(
               padding: const EdgeInsets.all(16.0),
@@ -190,20 +195,17 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
             ));
           }
 
-          // ( ✨ 已修改 ✨ ) 即使数据为空，也显示总评分卡
-          final List<OrderModel> orders = snapshot.data ?? [];
+          final List<ReviewModel> reviews = snapshot.data ?? []; // <-- 类型已更改
 
           return SingleChildScrollView(
             padding: const EdgeInsets.all(16.0),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                // 1. 显示总评分 (即使是 0)
-                _buildOverallRatingCard(orders),
+                _buildOverallRatingCard(reviews), // <-- 传入 reviews
                 const SizedBox(height: 24),
 
-                // 2. 仅当有评价时才显示列表
-                if (orders.isNotEmpty) ...[
+                if (reviews.isNotEmpty) ...[
                   const Text(
                     'All Reviews',
                     style: TextStyle(
@@ -213,13 +215,12 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
                     ),
                   ),
                   const SizedBox(height: 12),
-                  // 3. 显示评价列表
                   ListView.builder(
-                    itemCount: orders.length,
+                    itemCount: reviews.length,
                     shrinkWrap: true,
                     physics: const NeverScrollableScrollPhysics(),
                     itemBuilder: (context, index) {
-                      return _buildReviewCard(orders[index]);
+                      return _buildReviewCard(reviews[index]); // <-- 传入 review
                     },
                   ),
                 ],
