@@ -16,7 +16,9 @@ class MapPage extends StatefulWidget {
 
 class _MapPageState extends State<MapPage> {
   GoogleMapController? mapController;
-  LatLng _currentMapCenter = const LatLng(3.1390, 101.6869);
+  // --- ( ✨ UPDATED: Default location changed to New York, US ✨ ) ---
+  LatLng _currentMapCenter = const LatLng(40.7128, -74.0060); // New York City
+  // --- ( ✨ END UPDATE ✨ ) ---
   String _currentAddress = "Loading...";
   bool _locationReady = false;
 
@@ -31,26 +33,54 @@ class _MapPageState extends State<MapPage> {
   }
 
   Future<void> _getUserCurrentLocation() async {
-    final permission = await Geolocator.requestPermission();
-    if (permission == LocationPermission.denied || permission == LocationPermission.deniedForever) return;
+    try {
+      final permission = await Geolocator.requestPermission();
+      if (permission == LocationPermission.denied ||
+          permission == LocationPermission.deniedForever) {
+        // If permission is denied, just use the default US location
+        setState(() {
+          _locationReady = true;
+        });
+        await _getAddressFromLatLng(_currentMapCenter);
+        return;
+      }
 
-    final position = await Geolocator.getCurrentPosition(desiredAccuracy: LocationAccuracy.high);
-    final currentLatLng = LatLng(position.latitude, position.longitude);
+      final position = await Geolocator.getCurrentPosition(
+          desiredAccuracy: LocationAccuracy.high);
+      final currentLatLng = LatLng(position.latitude, position.longitude);
 
-    setState(() {
-      _currentMapCenter = currentLatLng;
-      _locationReady = true;
-    });
+      setState(() {
+        _currentMapCenter = currentLatLng;
+        _locationReady = true;
+      });
 
-    await _getAddressFromLatLng(currentLatLng);
+      await _getAddressFromLatLng(currentLatLng);
+      // --- ( ✨ NEW: Move map to user location AFTER getting it ✨ ) ---
+      mapController?.animateCamera(CameraUpdate.newLatLngZoom(currentLatLng, 17));
+      // --- ( ✨ END NEW ✨ ) ---
+    } catch (e) {
+      // Handle potential errors (like location services being off)
+      print("Error getting location: $e");
+      setState(() {
+        _locationReady = true; // Allow map to load with default US location
+      });
+      await _getAddressFromLatLng(_currentMapCenter); // Get address for default
+    }
   }
 
   Future<void> _getAddressFromLatLng(LatLng position) async {
     try {
-      final placemarks = await placemarkFromCoordinates(position.latitude, position.longitude);
-      final place = placemarks.first;
-      final address = "${place.street}, ${place.locality}, ${place.administrativeArea}, ${place.country}";
-      setState(() => _currentAddress = address.isEmpty ? "Unknown Location" : address);
+      final placemarks =
+          await placemarkFromCoordinates(position.latitude, position.longitude);
+      if (placemarks.isNotEmpty) {
+        final place = placemarks.first;
+        final address =
+            "${place.street}, ${place.locality}, ${place.administrativeArea}, ${place.country}";
+        setState(() =>
+            _currentAddress = address.isEmpty ? "Unknown Location" : address);
+      } else {
+        setState(() => _currentAddress = "Unknown Location");
+      }
     } catch (_) {
       setState(() => _currentAddress = "Unable to get address.");
     }
@@ -58,7 +88,10 @@ class _MapPageState extends State<MapPage> {
 
   void _onMapCreated(GoogleMapController controller) {
     mapController = controller;
-    mapController?.animateCamera(CameraUpdate.newLatLngZoom(_currentMapCenter, 17));
+    // --- ( ✨ UPDATED: Set initial position, but don't animate yet ✨ ) ---
+    // The animation will happen in _getUserCurrentLocation if successful
+    mapController?.moveCamera(CameraUpdate.newLatLngZoom(_currentMapCenter, 17));
+    // --- ( ✨ END UPDATE ✨ ) ---
   }
 
   void _onCameraMove(CameraPosition position) {
@@ -66,7 +99,10 @@ class _MapPageState extends State<MapPage> {
   }
 
   void _onCameraIdle() {
-    _getAddressFromLatLng(_currentMapCenter);
+    // To avoid spamming geocoding API, only update if not searching
+    if (_searchController.text.isEmpty) {
+      _getAddressFromLatLng(_currentMapCenter);
+    }
   }
 
   void _confirmLocation() {
@@ -84,7 +120,10 @@ class _MapPageState extends State<MapPage> {
       return;
     }
 
-    final url = 'https://maps.googleapis.com/maps/api/place/autocomplete/json?input=$input&key=${ApiConfig.googleMapsApiKey}&components=country:my';
+    // --- ( ✨ UPDATED: Changed country component to 'us' ✨ ) ---
+    final url =
+        'https://maps.googleapis.com/maps/api/place/autocomplete/json?input=$input&key=${ApiConfig.googleMapsApiKey}&components=country:us';
+    // --- ( ✨ END UPDATE ✨ ) ---
     final response = await http.get(Uri.parse(url));
     final data = json.decode(response.body);
 
@@ -96,34 +135,45 @@ class _MapPageState extends State<MapPage> {
   }
 
   void _selectPlace(String placeId) async {
-    final url = 'https://maps.googleapis.com/maps/api/place/details/json?place_id=$placeId&key=${ApiConfig.googleMapsApiKey}';
-    final response = await http.get(Uri.parse(url));
-    final data = json.decode(response.body);
-
-    final loc = data['result']['geometry']['location'];
-    final addr = data['result']['formatted_address'];
-    final newCenter = LatLng(loc['lat'], loc['lng']);
-
+    // Clear suggestions immediately for better UI
     setState(() {
-      _currentMapCenter = newCenter;
-      _currentAddress = addr;
-      _searchController.text = addr;
       _suggestions = [];
     });
 
-    mapController?.animateCamera(CameraUpdate.newLatLngZoom(newCenter, 17));
+    final url =
+        'https://maps.googleapis.com/maps/api/place/details/json?place_id=$placeId&key=${ApiConfig.googleMapsApiKey}';
+    final response = await http.get(Uri.parse(url));
+    final data = json.decode(response.body);
+
+    if (data['status'] == 'OK' && data['result'] != null) {
+      final loc = data['result']['geometry']['location'];
+      final addr = data['result']['formatted_address'];
+      final newCenter = LatLng(loc['lat'], loc['lng']);
+
+      setState(() {
+        _currentMapCenter = newCenter;
+        _currentAddress = addr;
+        _searchController.text =
+            addr; // Set text to full address
+      });
+
+      mapController?.animateCamera(CameraUpdate.newLatLngZoom(newCenter, 17));
+    }
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('Choose Your Location'), backgroundColor: kPrimaryActionColor),
+      appBar: AppBar(
+          title: const Text('Choose Your Location'),
+          backgroundColor: kPrimaryActionColor),
       body: !_locationReady
           ? const Center(child: CircularProgressIndicator())
           : Stack(
               children: [
                 GoogleMap(
-                  initialCameraPosition: CameraPosition(target: _currentMapCenter, zoom: 17.0),
+                  initialCameraPosition:
+                      CameraPosition(target: _currentMapCenter, zoom: 17.0),
                   myLocationEnabled: true,
                   myLocationButtonEnabled: true,
                   zoomControlsEnabled: false,
@@ -131,7 +181,9 @@ class _MapPageState extends State<MapPage> {
                   onCameraMove: _onCameraMove,
                   onCameraIdle: _onCameraIdle,
                 ),
-                const Center(child: Icon(Icons.location_on, size: 40, color: kPrimaryActionColor)),
+                const Center(
+                    child: Icon(Icons.location_on,
+                        size: 40, color: kPrimaryActionColor)),
                 Positioned(
                   top: 20,
                   left: 20,
@@ -139,16 +191,32 @@ class _MapPageState extends State<MapPage> {
                   child: Column(
                     children: [
                       Container(
-                        decoration: BoxDecoration(color: Colors.white, borderRadius: BorderRadius.circular(8), boxShadow: const [BoxShadow(color: Colors.black12, blurRadius: 4)]),
+                        decoration: BoxDecoration(
+                            color: Colors.white,
+                            borderRadius: BorderRadius.circular(8),
+                            boxShadow: const [
+                              BoxShadow(color: Colors.black12, blurRadius: 4)
+                            ]),
                         child: TextField(
                           controller: _searchController,
-                          decoration: const InputDecoration(hintText: 'Search location', prefixIcon: Icon(Icons.search), border: InputBorder.none, contentPadding: EdgeInsets.all(12)),
+                          decoration: const InputDecoration(
+                              hintText: 'Search location',
+                              prefixIcon: Icon(Icons.search),
+                              border: InputBorder.none,
+                              contentPadding: EdgeInsets.all(12)),
                         ),
                       ),
                       if (_suggestions.isNotEmpty)
                         Container(
+                          height:
+                              200, // Limit height of suggestions
                           margin: const EdgeInsets.only(top: 4),
-                          decoration: BoxDecoration(color: Colors.white, borderRadius: BorderRadius.circular(8), boxShadow: const [BoxShadow(color: Colors.black12, blurRadius: 4)]),
+                          decoration: BoxDecoration(
+                              color: Colors.white,
+                              borderRadius: BorderRadius.circular(8),
+                              boxShadow: const [
+                                BoxShadow(color: Colors.black12, blurRadius: 4)
+                              ]),
                           child: ListView.builder(
                             shrinkWrap: true,
                             itemCount: _suggestions.length,
@@ -168,10 +236,21 @@ class _MapPageState extends State<MapPage> {
                 Align(
                   alignment: Alignment.topCenter,
                   child: Container(
-                    margin: const EdgeInsets.only(top: 100, left: 20, right: 20),
+                    margin: const EdgeInsets.only(
+                        top: 90,
+                        left: 20,
+                        right: 20), // Moved down to avoid search
                     padding: const EdgeInsets.all(12),
-                    decoration: BoxDecoration(color: kCardColor, borderRadius: BorderRadius.circular(8), boxShadow: const [BoxShadow(color: Colors.black12, blurRadius: 4)]),
-                    child: Text(_currentAddress, textAlign: TextAlign.center, style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 14)),
+                    decoration: BoxDecoration(
+                        color: kCardColor,
+                        borderRadius: BorderRadius.circular(8),
+                        boxShadow: const [
+                          BoxShadow(color: Colors.black12, blurRadius: 4)
+                        ]),
+                    child: Text(_currentAddress,
+                        textAlign: TextAlign.center,
+                        style: const TextStyle(
+                            fontWeight: FontWeight.bold, fontSize: 14)),
                   ),
                 ),
                 Align(
@@ -182,11 +261,17 @@ class _MapPageState extends State<MapPage> {
                       onPressed: _confirmLocation,
                       style: ElevatedButton.styleFrom(
                         backgroundColor: kPrimaryActionColor,
-                        padding: const EdgeInsets.symmetric(horizontal: 60, vertical: 18),
-                        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+                        padding: const EdgeInsets.symmetric(
+                            horizontal: 60, vertical: 18),
+                        shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(12)),
                         elevation: 4,
                       ),
-                      child: const Text('Confirm Location', style: TextStyle(fontSize: 20, color: kCardColor, fontWeight: FontWeight.bold)),
+                      child: const Text('Confirm Location',
+                          style: TextStyle(
+                              fontSize: 20,
+                              color: kCardColor,
+                              fontWeight: FontWeight.bold)),
                     ),
                   ),
                 ),
