@@ -6,8 +6,6 @@ import 'package:foodiebox/screens/users/order_tracking_page.dart';
 import 'package:foodiebox/util/styles.dart';
 import 'package:foodiebox/widgets/base_page.dart';
 import 'package:intl/intl.dart';
-import 'pickup_confirmation_page.dart';
-import 'order_history_detail_page.dart';
 
 class OrdersPage extends StatefulWidget {
   const OrdersPage({super.key});
@@ -75,24 +73,17 @@ class _OrdersPageState extends State<OrdersPage>
 
   Widget _buildOrderList(BuildContext context,
       {required bool isOngoing, String? userId}) {
-    // --- ( ✨ UPDATED STATUSES - CASE INSENSITIVE FIX ✨ ) ---
-    // We now check for both lowercase and uppercase versions
+    // --- FIX: Add new pickup status to ongoing list ---
     final ongoingStatuses = [
-      'received', 'Received',
-      'Preparing', 'preparing',
-      'Prepared', 'prepared',
-      'Ready for Pickup', 'ready for pickup',
-      'paid_pending_pickup',
-      'Delivering', 'delivering',
+      'received',
+      'preparing',
+      'delivering',
+      'pending',
+      'paid_pending_pickup' // <-- ADDED
     ];
-    
-    final historyStatuses = [
-      'completed', 'Completed',
-      'cancelled', 'Cancelled', // <-- This is the important fix
-      'Delivered', 'delivered',
-      'Picked Up', 'picked up'
-    ];
-    // --- ( ✨ END UPDATED STATUSES ✨ ) ---
+    // --- END FIX ---
+
+    final historyStatuses = ['completed', 'cancelled'];
 
     if (userId == null) {
       return const Center(
@@ -116,24 +107,8 @@ class _OrdersPageState extends State<OrdersPage>
         }
 
         if (snapshot.hasError) {
-          // --- ADDED: Show the index error clearly ---
-          if (snapshot.error.toString().contains('FAILED_PRECONDITION')) {
-            return const Center(
-              child: Padding(
-                padding: EdgeInsets.all(20.0),
-                child: Text(
-                  'Database Index Required. Please create the index in your Firebase console (check the error log for the link).',
-                  textAlign: TextAlign.center,
-                  style:
-                      TextStyle(color: Colors.red, fontWeight: FontWeight.bold),
-                ),
-              ),
-            );
-          }
-          // --- END ADDED ---
           return Center(
-              child: Text('An error occurred: ${snapshot.error}',
-                  style: kHintTextStyle));
+              child: Text('An error occurred.', style: kHintTextStyle));
         }
 
         if (!snapshot.hasData || snapshot.data!.docs.isEmpty) {
@@ -157,56 +132,26 @@ class _OrdersPageState extends State<OrdersPage>
     );
   }
 
-
   Widget _buildOrderCard(BuildContext context, OrderModel order,
       {required bool isOngoing}) {
-    // --- UPDATED: Use the status string directly, handle capitalization ---
-    final String statusText = order.status.isNotEmpty
-        ? order.status[0].toUpperCase() + order.status.substring(1)
-        : 'Unknown';
-    // --- END UPDATED ---
-
+    final String statusText =
+        order.status[0].toUpperCase() + order.status.substring(1);
     final TextStyle statusStyle = _getStatusStyle(order.status);
     final String formattedDateTime = _formatDateTime(order.timestamp.toDate());
 
     return GestureDetector(
-      // --- ( ✨ UPDATED NAVIGATION LOGIC ✨ ) ---
+      // --- FIX: Only allow tap to track for delivery orders ---
       onTap: () {
-        if (isOngoing) {
-          if (order.orderType == 'Delivery') {
-            // Ongoing Delivery -> Track Page
-            Navigator.push(
-              context,
-              MaterialPageRoute(
-                builder: (_) => OrderTrackingPage(orderId: order.id),
-              ),
-            );
-          } else if (order.orderType == 'Pickup') {
-            // Ongoing Pickup -> QR Code Page
-            Navigator.push(
-              context,
-              MaterialPageRoute(
-                builder: (_) => PickupConfirmationPage(
-                  orderId: order.id,
-                  pickupId: order.pickupId ?? 'NA',
-                  vendorName: order.vendorName ?? 'Store',
-                  vendorAddress: order.vendorAddress ?? 'No address',
-                  total: order.total,
-                ),
-              ),
-            );
-          }
-        } else {
-          // History Order (Any type) -> Details Page
+        if (isOngoing && order.orderType == 'Delivery') {
           Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (_) => OrderHistoryDetailsPage(order: order),
+              builder: (_) => OrderTrackingPage(orderId: order.id),
             ),
           );
         }
       },
-      // --- ( ✨ END UPDATED NAVIGATION LOGIC ✨ ) ---
+      // --- END FIX ---
       child: Container(
         margin: const EdgeInsets.only(bottom: 16),
         padding: const EdgeInsets.all(12),
@@ -229,7 +174,7 @@ class _OrdersPageState extends State<OrdersPage>
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
                 Expanded(
-                  child: Text(order.vendorName ?? 'Store',
+                  child: Text(order.vendorName as String,
                       style: kLabelTextStyle.copyWith(fontSize: 16),
                       overflow: TextOverflow.ellipsis),
                 ),
@@ -248,7 +193,7 @@ class _OrdersPageState extends State<OrdersPage>
             const SizedBox(height: 8),
             Text(formattedDateTime,
                 style: kHintTextStyle.copyWith(fontSize: 13)),
-            
+
             // --- FIX: Check if address is null before showing it ---
             if (order.address != null && order.address!.isNotEmpty)
               Padding(
@@ -258,18 +203,7 @@ class _OrdersPageState extends State<OrdersPage>
                     maxLines: 1,
                     overflow: TextOverflow.ellipsis),
               ),
-
-            // --- ADDED: Show Pickup ID if it's a pickup order ---
-            if (order.orderType == 'Pickup' && order.pickupId != null)
-              Padding(
-                padding: const EdgeInsets.only(top: 4.0),
-                child: Text('Pickup ID: ${order.pickupId!}',
-                    style: kHintTextStyle.copyWith(
-                        fontSize: 13,
-                        color: kPrimaryActionColor,
-                        fontWeight: FontWeight.bold)),
-              ),
-            // --- END ADDED ---
+            // --- END FIX ---
 
             const SizedBox(height: 8),
             Row(
@@ -277,15 +211,18 @@ class _OrdersPageState extends State<OrdersPage>
               children: [
                 Text('Total: RM${order.total.toStringAsFixed(2)}',
                     style: kLabelTextStyle.copyWith(fontSize: 15)),
-                
+
                 // --- FIX: Show correct icon for delivery or pickup ---
-                if (order.orderType == 'Delivery' && order.driverId != null && isOngoing)
+                if (order.orderType == 'Delivery' &&
+                    order.driverId != null &&
+                    isOngoing)
                   const Icon(Icons.local_shipping,
                       color: kPrimaryActionColor, size: 20),
 
                 if (order.orderType == 'Pickup' && isOngoing)
                   const Icon(Icons.store, // Icon for pickup
-                      color: kPrimaryActionColor, size: 20),
+                      color: kPrimaryActionColor,
+                      size: 20),
                 // --- END FIX ---
               ],
             ),
@@ -295,15 +232,10 @@ class _OrdersPageState extends State<OrdersPage>
     );
   }
 
-
   TextStyle _getStatusStyle(String status) {
     Color color;
-    // --- ( ✨ UPDATED STATUS STYLES ✨ ) ---
-    // This switch already uses .toLowerCase(), so it's safe!
-    switch (status.toLowerCase()) {
+    switch (status) {
       case 'completed':
-      case 'delivered':
-      case 'picked up':
         color = Colors.green;
         break;
       case 'cancelled':
@@ -313,20 +245,19 @@ class _OrdersPageState extends State<OrdersPage>
         color = Colors.orange;
         break;
       case 'preparing':
-      case 'prepared':
         color = Colors.amber;
         break;
-      case 'ready for pickup':
+      // --- FIX: Add new status color ---
       case 'paid_pending_pickup':
         color = Colors.blue;
         break;
-      case 'Received':
+      // --- END FIX ---
+      case 'received':
       case 'pending':
       default:
         color = kPrimaryActionColor;
         break;
     }
-    // --- ( ✨ END UPDATED STATUS STYLES ✨ ) ---
     return TextStyle(fontSize: 13, fontWeight: FontWeight.w500, color: color);
   }
 
